Possible terminating byte fix
There is an issue with the server receiving a different terminating byte
than is sent by the client. The protocol is to send 0xFF as an end of
message, but through sending, it is received on the server end as 0xEF.
So, I am looking for 0xEF instead of 0xFF to end the message.

The client is using the Java PrintWriter to send 0xFFFD as an integer.
0xFFFD is the UTF-8 encoding of 0xFF.